fix: update cdn urls
